Allow the -p parameter to contain decimals in order to allow polling times less than one second.

+++ src/mem_tracker.h
@@ -7,7 +7,7 @@ void *mem_tracker(void *arg);
 
 struct tracker_arg{
     print_func_t print_func;
-    unsigned int poll;
+    unsigned long poll;
     double print_threshold;
 };
